moving static.py
