+++ museum_of_dreams_project/settings/aws.py
@@ -1,5 +1,4 @@
 from .base import *
-import museum_of_dreams_project.settings.static as static
 
 
 SECRET_KEY = os.environ.get("SECRET_KEY")
