+++ museum_of_dreams_project/static.py
@@ -1,5 +1,5 @@
 from storages.backends.s3boto3 import S3Boto3Storage
-from .aws import STATIC_ROOT, MEDIA_ROOT
+from .settings.aws import STATIC_ROOT, MEDIA_ROOT
 
 
 class StaticStorage(S3Boto3Storage):
